+++ src/main/java/org/morphix/lang/Case.java
@@ -30,6 +30,9 @@ public enum Case {
 	 * "lowerCamelCase").
 	 */
 	LOWER_CAMEL {
+		/**
+		 * @see Case#format(String[], Locale)
+		 */
 		@Override
 		public String format(final String[] words, final Locale locale) {
 			if (words.length == 0) {
@@ -37,7 +40,7 @@ public enum Case {
 			}
 			StringBuilder sb = new StringBuilder(words[0].toLowerCase(locale));
 			for (int i = 1; i < words.length; ++i) {
-				sb.append(capitalize(words[i]));
+				sb.append(capitalize(words[i], locale));
 			}
 			return sb.toString();
 		}
@@ -47,11 +50,14 @@ public enum Case {
 	 * Upper camel case style, where all words are capitalized (e.g. "UpperCamelCase").
 	 */
 	UPPER_CAMEL {
+		/**
+		 * @see Case#format(String[], Locale)
+		 */
 		@Override
 		public String format(final String[] words, final Locale locale) {
 			StringBuilder sb = new StringBuilder();
 			for (String word : words) {
-				sb.append(capitalize(word));
+				sb.append(capitalize(word, locale));
 			}
 			return sb.toString();
 		}
@@ -61,9 +67,25 @@ public enum Case {
 	 * Snake case style, where all words are in lowercase and separated by underscores (e.g. "snake_case").
 	 */
 	SNAKE {
+		/**
+		 * The separator used for joining words in snake case style, which is an underscore ("_").
+		 */
+		public static final String SEPARATOR = "_";
+
+		/**
+		 * @see Case#format(String[], Locale)
+		 */
 		@Override
 		public String format(final String[] words, final Locale locale) {
-			return join(words, "_", locale, Letter.LOWER);
+			return join(words, SEPARATOR, locale, Letter.LOWER);
+		}
+
+		/**
+		 * @see Case#wordSeparator()
+		 */
+		@Override
+		public String wordSeparator() {
+			return SEPARATOR;
 		}
 	},
 
@@ -71,9 +93,25 @@ public enum Case {
 	 * Upper snake case style, where all words are in uppercase and separated by underscores (e.g. "UPPER_SNAKE_CASE").
 	 */
 	UPPER_SNAKE {
+		/**
+		 * The separator used for joining words in upper snake case style, which is an underscore ("_").
+		 */
+		public static final String SEPARATOR = "_";
+
+		/**
+		 * @see Case#format(String[], Locale)
+		 */
 		@Override
 		public String format(final String[] words, final Locale locale) {
-			return join(words, "_", locale, Letter.UPPER);
+			return join(words, SEPARATOR, locale, Letter.UPPER);
+		}
+
+		/**
+		 * @see Case#wordSeparator()
+		 */
+		@Override
+		public String wordSeparator() {
+			return SEPARATOR;
 		}
 	},
 
@@ -81,9 +119,25 @@ public enum Case {
 	 * Kebab case style, where all words are in lowercase and separated by hyphens (e.g. "kebab-case").
 	 */
 	KEBAB {
+		/**
+		 * The separator used for joining words in kebab case style, which is a hyphen ("-").
+		 */
+		public static final String SEPARATOR = "-";
+
+		/**
+		 * @see Case#format(String[], Locale)
+		 */
 		@Override
 		public String format(final String[] words, final Locale locale) {
-			return join(words, "-", locale, Letter.LOWER);
+			return join(words, SEPARATOR, locale, Letter.LOWER);
+		}
+
+		/**
+		 * @see Case#wordSeparator()
+		 */
+		@Override
+		public String wordSeparator() {
+			return SEPARATOR;
 		}
 	};
 
@@ -152,6 +206,17 @@ public enum Case {
 	public abstract String format(String[] words, Locale locale);
 
 	/**
+	 * Returns the separator used for joining words in this case style. By default, it returns an empty string, but it can
+	 * be overridden by specific case styles that use a separator (e.g. SNAKE, UPPER_SNAKE, KEBAB) to return the appropriate
+	 * separator (e.g. "_" for snake case and "-" for kebab case).
+	 *
+	 * @return the separator used for joining words in this case style
+	 */
+	public String wordSeparator() {
+		return "";
+	}
+
+	/**
 	 * Tokenizes the input name into an array of words by splitting on hyphens, underscores, and camel case boundaries. The
 	 * method first normalizes the input by replacing hyphens with underscores, then splits the normalized string on
 	 * underscores, and finally further splits each part on camel case boundaries. The resulting tokens are filtered to
@@ -170,21 +235,27 @@ public enum Case {
 		String[] words = new String[(length / 2 + 1) + 1];
 		int wordCount = 0;
 
-		int start = -1;
+		int wordStartIndex = -1;
 		for (int i = 0; i < length; ++i) {
 			char c = name.charAt(i);
-			if (c == '_' || c == '-') {
-				if (start >= 0) {
-					words[wordCount] = name.substring(start, i);
+			boolean isWordSeparator = c == '_' || c == '-';
+			if (isWordSeparator) {
+				if (wordStartIndex >= 0) {
+					words[wordCount] = name.substring(wordStartIndex, i);
 					++wordCount;
-					start = -1;
+					wordStartIndex = -1;
 				}
+			} else if (wordStartIndex < 0) {
+				wordStartIndex = i;
+				isWordSeparator = true;
+			}
+			if (isWordSeparator) {
 				continue;
 			}
-			if (start < 0) {
-				start = i;
-				continue;
-			}
+			// this part is an optimization to avoid checking for camel case boundaries on every character, we only check when the
+			// previous character is a letter or a digit and the current character is a letter, or when the previous character is an
+			// uppercase letter and the current character is an uppercase letter followed by a lowercase letter (e.g. "JSONValue"
+			// should be split into "JSON" and "Value")
 			char prev = name.charAt(i - 1);
 
 			boolean wordEnded = (Character.isLowerCase(prev) && Character.isUpperCase(c))
@@ -193,12 +264,12 @@ public enum Case {
 							&& i + 1 < length && Character.isLowerCase(name.charAt(i + 1)));
 
 			if (wordEnded) {
-				words[wordCount] = name.substring(start, i);
+				words[wordCount] = name.substring(wordStartIndex, i);
 				++wordCount;
-				start = i;
+				wordStartIndex = i;
 			}
 		}
-		words[wordCount] = name.substring(start);
+		words[wordCount] = name.substring(wordStartIndex);
 		++wordCount;
 		return Arrays.copyOf(words, wordCount);
 	}
@@ -230,10 +301,26 @@ public enum Case {
 	 * @return the capitalized word with the first letter in uppercase and the rest in lowercase
 	 */
 	public static String capitalize(final String word) {
+		return capitalize(word, Locale.ROOT);
+	}
+
+	/**
+	 * Capitalizes the first letter of the given word and converts the rest of the letters to lowercase. If the input word
+	 * is empty, it returns the empty string. This method is used to convert words to a format suitable for camel case
+	 * styles.
+	 *
+	 * @param word the input word to capitalize
+	 * @param locale the locale to use for any locale-specific formatting
+	 * @return the capitalized word with the first letter in uppercase and the rest in lowercase
+	 */
+	public static String capitalize(final String word, final Locale locale) {
 		if (null == word || word.isEmpty()) {
 			return word;
 		}
-		return Character.toUpperCase(word.charAt(0)) +
-				word.substring(1).toLowerCase(Locale.ROOT);
+		if (word.length() == 1) {
+			return word.toUpperCase(locale);
+		}
+		return word.substring(0, 1).toUpperCase(locale) +
+				word.substring(1).toLowerCase(locale);
 	}
 }